chore: rename useToggleExpandHelper file
