+++ meteor/client/ui/Settings/BlueprintConfigSchema/index.tsx
@@ -12,7 +12,7 @@ import { useOverrideOpHelper, WrappedOverridableItemNormal } from '../util/Overr
 import { JSONSchema } from '@sofie-automation/shared-lib/dist/lib/JSONSchemaTypes'
 import deepmerge from 'deepmerge'
 import { SchemaFormSofieEnumDefinition, translateStringIfHasNamespaces } from '../../../lib/forms/schemaFormUtil'
-import { useToggleExpandHelper } from '../util/ToggleExpandedHelper'
+import { useToggleExpandHelper } from '../../util/useToggleExpandHelper'
 import { SourceLayers } from '@sofie-automation/corelib/dist/dataModel/ShowStyleBase'
 import { ConfigCategoryEntry } from './CategoryEntry'
